@@ -1,20 +1,22 @@
 // @ts-check
-// Note: type annotations allow type checking and IDEs autocompletion
+// `@type` JSDoc annotations allow editor autocompletion and type checking
+// (when paired with `@ts-check`).
+// There are various equivalent ways to declare your Docusaurus config.
+// See: https://docusaurus.io/docs/api/docusaurus-config
 
-const lightCodeTheme = require("prism-react-renderer/themes/github");
-const darkCodeTheme = require("prism-react-renderer/themes/dracula");
+import { themes as prismThemes } from "prism-react-renderer";
 
-/** @type {import("@docusaurus/types").Config} */
-module.exports = {
+/** @type {import('@docusaurus/types').Config} */
+const config = {
   title: "ilya Zuev",
   tagline: "more than a Software Developer",
   favicon: "img/photo.png",
 
   // Set the production url of your site here
-  url: "https://ilya.zuev.eu", // https://ilyazuev.github.io
+  url: "https://ilya.zuev.eu",
   // Set the /<baseUrl>/ pathname under which your site is served
   // For GitHub pages deployment, it is often '/<projectName>/'
-  baseUrl: "/", // /cv/
+  baseUrl: "/",
 
   // GitHub pages deployment config.
   // If you aren't using GitHub pages, you don't need these.
@@ -24,9 +26,9 @@ module.exports = {
   onBrokenLinks: "throw",
   onBrokenMarkdownLinks: "warn",
 
-  // Even if you don't use internalization, you can use this field to set useful
-  // metadata like html lang. For example, if your site is Chinese, you may want
-  // to replace "en" with "zh-Hans".
+  // Even if you don't use internationalization, you can use this field to set
+  // useful metadata like html lang. For example, if your site is Chinese, you
+  // may want to replace "en" with "zh-Hans".
   i18n: {
     defaultLocale: "en",
     locales: ["en"],
@@ -35,25 +37,27 @@ module.exports = {
   presets: [
     [
       "classic",
-      /** @type {import("@docusaurus/preset-classic").Options} */
+      /** @type {import('@docusaurus/preset-classic').Options} */
       ({
         docs: {
-          sidebarPath: require.resolve("./sidebars.js"),
+          sidebarPath: "./sidebars.js",
           routeBasePath: "/",
         },
         blog: false,
         theme: {
-          customCss: require.resolve("./src/css/custom.css"),
+          customCss: "./src/css/custom.css",
         },
       }),
     ],
   ],
 
   themeConfig:
-    /** @type {import("@docusaurus/preset-classic").ThemeConfig} */
-    {
+    /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
+    ({
+      // Replace with your project's social card
+      image: "img/docusaurus-social-card.jpg",
       navbar: {
-        title: "ilya Zuev", // logo: { alt: "logo", src: "img/logo.svg", },
+        title: "ilya Zuev", // logo: { alt: "My Site Logo", src: "img/logo.svg", },
         items: [
           {
             type: "docSidebar",
@@ -88,8 +92,10 @@ module.exports = {
         ],
       },
       prism: {
-        theme: lightCodeTheme,
-        darkTheme: darkCodeTheme,
+        theme: prismThemes.github,
+        darkTheme: prismThemes.dracula,
       },
-    },
+    }),
 };
+
+export default config;
