@@ -10,4 +10,4 @@ const sidebars = {
   ],
 };
 
-module.exports = sidebars;
+export default sidebars; //module.exports = sidebars;
